+++ lib/backends/webgl/op-resolve-rules.ts
@@ -22,7 +22,7 @@ import {WebGLPad} from './ops/pad';
 import {WebGLAveragePool, WebGLGlobalAveragePool, WebGLGlobalMaxPool, WebGLMaxPool} from './ops/pool';
 import * as reduceOps from './ops/reduce';
 import {WebGLReshape} from './ops/reshape';
-import {WebGLResizePacked} from './ops/resize-bilinear-packed';
+import {WebGLResizePacked} from './ops/resize-packed';
 import {WebGLSlice, WebGLSliceV10} from './ops/slice';
 import {WebGLSoftmax} from './ops/softmax';
 import {WebGLSplit} from './ops/split';
@@ -86,8 +86,8 @@ export const WEBGL_OP_RESOLVE_RULES: ReadonlyArray<OpSet.ResolveRule> = [
   ['Relu', '', '6+', () => new unaryOps.WebGLUnaryOp(FLOAT_TYPES, unaryOps.glslRelu())],
   ['Reshape', '', '5+', () => new WebGLReshape()],
   // ['ResizeBilinearPacked', '', '1+', () => new WebGLResizePacked()],
-  ['ResizePacked', '', '10', () => new WebGLResizePacked(10)],
-  ['ResizePacked', '', '11+', () => new WebGLResizePacked(11)],
+  ['Resize', '', '10', () => new WebGLResizePacked(10)],
+  ['Resize', '', '11+', () => new WebGLResizePacked(11)],
   ['Sigmoid', '', '6+', () => new unaryOps.WebGLUnaryOp(FLOAT_TYPES, unaryOps.glslSigmoid())],
   ['Sin', '', '7+', () => new unaryOps.WebGLUnaryOp(FLOAT_TYPES, unaryOps.glslSin())],
   ['Slice', '', '10+', () => new WebGLSliceV10()],  // TODO: support 'steps' for Slice-10
