--- test/data/ops/resize_bilinear_pack.jsonc
@@ -1,46 +1,0 @@
-[
-  {
-    "name": "ResizeBilinearPacked with '0' and '-1' in the shape tensor input",
-    "operator": "ResizePacked",
-    "opsets": [{"domain": "", "version": "11"}],
-    "attributes": [
-      // { "name": "scales", "data": [1.0, 1.0, 2.0, 3.0], "type": "floats" },
-        {"name": "coordinate_transformation_mode", "data": "asymmetric", "type": "string"},
-        {"name": "mode", "data": "linear", "type": "string"}
-    ],
-    "cases": [
-      {
-        "name": "T[0,3]",
-        "inputs": [
-          {
-            "data": [2, 2, 4, 4],
-            "dims": [1, 1, 2, 2],
-            "type": "float32"
-          },
-          {
-            "data": [],
-            "dims": [0],
-            "type": "float32"
-          },
-          {
-            "data": [],
-            "dims": [0],
-            "type": "float32"
-          },
-          {
-            "data": [1, 1, 3, 3],
-            "dims": [4],
-            "type": "int32"
-          }
-        ],
-        "outputs": [
-          {
-            "data": [2, 2, 2, 3.33333, 3.33333, 3.33333, 4, 4, 4],
-            "dims": [1, 1, 3, 3],
-            "type": "float32"
-          }
-        ]
-      }
-    ]
-  }
-]
