+++ test/unittests/backends/webgl/test_pack_unpack.ts
@@ -6,7 +6,7 @@ import {Attribute} from '../../../../lib/attribute';
 import {Backend, InferenceHandler, SessionHandler} from '../../../../lib/backend';
 import {WebGLInferenceHandler} from '../../../../lib/backends/webgl/inference-handler';
 import {WebGLPack} from '../../../../lib/backends/webgl/ops/pack';
-import {WebGLResizePacked} from '../../../../lib/backends/webgl/ops/resize-bilinear-packed';
+import {WebGLResizePacked} from '../../../../lib/backends/webgl/ops/resize-packed';
 import {WebGLUnpack} from '../../../../lib/backends/webgl/ops/unpack';
 import {Profiler} from '../../../../lib/instrument';
 import {Tensor} from '../../../../lib/tensor';
